Address the fifty-ninth nearby review round
- The presence bookkeeping moves out of the companion service. That class
  extends android.companion.CompanionDeviceService, which does not exist
  before API 31, and AndroidNearbyBackend -- which every transport,
  ranging or companion build constructs, from API 21 up -- referenced it
  in its constructor. The class failed to resolve on anything older and
  the NoClassDefFoundError escaped into the reflective catch that builds
  the bridge, so the whole nearby stack reported itself unsupported on
  Android 21 to 30, where transport and ranging work perfectly well.

  It lives in NearbyPresenceStore now, which touches nothing newer than
  SharedPreferences. With the coupling gone the service is also deleted
  from a build that never observes presence, which it could not be before.

- Android accessory ranging answers NOT_SUPPORTED, as RangingSession
  .startAccessory and the developer guide have both documented all along.
  The method exists for Apple's accessory handshake and Android has none:
  what an accessory publishes there is a vendor format naming a channel
  and a session, which the app parses into a RangingToken for start().
  Reading those bytes AS a token was worse than refusing them -- real
  accessory data answered INVALID_TOKEN, and only an app passing a token
  it should have given to start() succeeded, so cross-platform code
  branching on the documented error never saw it. The capability bit goes
  with it: the port, the javadoc and the guide now say one thing.
